рефакторинг кода и структуры проекта и подготовка для использования в виде пакета, запуск в режиме сервера в docker-контейнере
