--- auth.yaml
@@ -1,15 +1,0 @@
-auth:
-    github:
-        token: TOKEN
-    email:
-        login: LOGIN
-        password: PASSWORD
-        return-address: "@gmail.com"
-        imap:
-            server: imap.gmail.com
-            port: 993
-        smtp:
-            server: smtp.gmail.com
-            port: 465
-    google:
-        credentials-file: credentials.json
