+++ static/app/auth.yaml
@@ -1,0 +1,18 @@
+requests_timeout: 10
+auth:
+  moss_userid: "userid"
+  github:
+    token: "token"
+  appveyor:
+    account: "account"
+    token: "token"
+  email:
+     login: "login"
+     password: "pwd"
+     return-address: "@gmail.com"
+       imap:
+         server: imap.gmail.com
+         port: 993
+       smtp:
+         server: smtp.gmail.com
+         port: 465
