+++ static/app/logging.yaml
