+++ static/courses/example-course-config.yml
@@ -1,0 +1,53 @@
+course:
+  name: Operating systems
+  alt_names:
+    - "OS"
+    - "Операционные системы"
+    - "ОС"
+  semester: Spring 2021
+  email: k43guap@ya.ru
+  timezone: MSK
+  github:
+    organization: suai-os-2021
+    teacher_github_logins:
+      - "Mark Polyak"
+      - "markpolyak"
+  google:
+    spreadsheet: 1DPTMvkkWY6Tm32K0vXp1kDWJTTPFeHgBGHKAALdjvpY
+    info-sheet: График
+    task-id-column: 0
+    student-name-column: 1
+    lab-column-offset: 1
+
+  labs:
+    - id: 0
+      github-prefix: os-task0
+      short-name: ЛР0
+      taskid_shift: 4
+      taskid-max: 20
+      penalty-max: 6
+        # ignore-task-id: True
+      ci:
+        - workflows
+      files:
+        - typescript
+      moss:
+        language: c
+        max_matches: 100
+
+    - id: 1
+      github-prefix: os-task1
+      short-name: ЛР1
+      taskid-max: 20
+      penalty-max: 6
+      ci:
+        - travis
+        - workflows
+      files:
+        - lab1.sh
+      moss:
+        language: c
+        max-matches: 1000
+        basefiles:
+          - repo: k43guap/os-course-task2
+            filename: lab1.sh
