+++ static/courses/trash/operating-systems-2020.yaml
@@ -1,5 +1,5 @@
 ---
-# Configuration file for lab-grader
+# Configuration file for lab_grader
 
 course:
     name: Operating systems
@@ -7,42 +7,24 @@ course:
         - OS
         - Операционные системы
         - ОС
-    semester: Spring 2021
+    semester: Spring 2020
     email: k43guap@ya.ru
-    timezone: MSK
     github:
-        organization: suai-os-2021
+        organization: suai-os-2020
         teachers:
             - "Mark Polyak"
             - markpolyak
     google:
-        spreadsheet: 1DPTMvkkWY6Tm32K0vXp1kDWJTTPFeHgBGHKAALdjvpY
-        info-sheet: График
-        task-id-column: 0
-        student-name-column: 1
-        lab-column-offset: 1
+        spreadsheet: 1ymyU98eB0HYUzVTgrArbtEOkiU3lnKSOS6BUNkssbTE
+        info-sheet: План
     labs:
-        "0":
-            github-prefix: os-task0
-            short-name: ЛР0
+        "1":
+            github-prefix: os-task1
             taskid-max: 20
             penalty-max: 6
             # ignore-task-id: True
             ci:
-                - workflows
-            files:
-                - typescript
-            moss:
-                language: c
-                max-matches: 1000
-        "1":
-            github-prefix: os-task1
-            short-name: ЛР1
-            taskid-max: 20
-            penalty-max: 6
-            ci:
                 - travis
-                - workflows
             files:
                 - lab1.sh
             moss:
@@ -54,12 +36,11 @@ course:
                         filename: lab1.sh
         "2":
             github-prefix: os-task2
-            short-name: ЛР2
             taskid-max: 20
             taskid-shift: 4
             penalty-max: 8
             ci:
-                - workflows
+                - travis
             files:
                 - lab2.cpp
             moss:
@@ -74,11 +55,10 @@ course:
                         filename: examples/ex3.cpp
         "3":
             github-prefix: os-task3
-            short-name: ЛР3
             taskid-max: 20
             penalty-max: 8
             ci:
-                - workflows
+                - appveyor
             files:
                 - lab3.cpp
             moss:
@@ -90,11 +70,10 @@ course:
                         filename: lab3.cpp
         "4":
             github-prefix: os-task4
-            short-name: ЛР4
             taskid-max: 30
             penalty-max: 10
             ci:
-                - workflows
+                - travis
             files:
                 - lab4.cpp
             moss:
@@ -106,13 +85,11 @@ course:
                         filename: lab4.cpp
         "5":
             github-prefix: os-task5
-            short-name: ЛР5
             taskid-max: 30
             penalty-max: 8
             ignore-completion-date: True
         "6":
             github-prefix: os-task5
-            short-name: ЛР6
             taskid-max: 30
             penalty-max: 7
