+++ static/courses/trash/operating-systems-2021.yaml
@@ -1,5 +1,5 @@
 ---
-# Configuration file for lab-grader
+# Configuration file for lab_grader
 
 course:
     name: Operating systems
@@ -7,24 +7,42 @@ course:
         - OS
         - Операционные системы
         - ОС
-    semester: Spring 2020
+    semester: Spring 2021
     email: k43guap@ya.ru
+    timezone: MSK
     github:
-        organization: suai-os-2020
+        organization: suai-os-2021
         teachers:
             - "Mark Polyak"
             - markpolyak
     google:
-        spreadsheet: 1ymyU98eB0HYUzVTgrArbtEOkiU3lnKSOS6BUNkssbTE
-        info-sheet: План
+        spreadsheet: 1DPTMvkkWY6Tm32K0vXp1kDWJTTPFeHgBGHKAALdjvpY
+        info-sheet: График
+        task-id-column: 0
+        student-name-column: 1
+        lab-column-offset: 1
     labs:
-        "1":
-            github-prefix: os-task1
+        "0":
+            github-prefix: os-task0
+            short-name: ЛР0
             taskid-max: 20
             penalty-max: 6
             # ignore-task-id: True
             ci:
+                - workflows
+            files:
+                - typescript
+            moss:
+                language: c
+                max-matches: 1000
+        "1":
+            github-prefix: os-task1
+            short-name: ЛР1
+            taskid-max: 20
+            penalty-max: 6
+            ci:
                 - travis
+                - workflows
             files:
                 - lab1.sh
             moss:
@@ -36,11 +54,12 @@ course:
                         filename: lab1.sh
         "2":
             github-prefix: os-task2
+            short-name: ЛР2
             taskid-max: 20
             taskid-shift: 4
             penalty-max: 8
             ci:
-                - travis
+                - workflows
             files:
                 - lab2.cpp
             moss:
@@ -55,10 +74,11 @@ course:
                         filename: examples/ex3.cpp
         "3":
             github-prefix: os-task3
+            short-name: ЛР3
             taskid-max: 20
             penalty-max: 8
             ci:
-                - appveyor
+                - workflows
             files:
                 - lab3.cpp
             moss:
@@ -70,10 +90,11 @@ course:
                         filename: lab3.cpp
         "4":
             github-prefix: os-task4
+            short-name: ЛР4
             taskid-max: 30
             penalty-max: 10
             ci:
-                - travis
+                - workflows
             files:
                 - lab4.cpp
             moss:
@@ -85,11 +106,13 @@ course:
                         filename: lab4.cpp
         "5":
             github-prefix: os-task5
+            short-name: ЛР5
             taskid-max: 30
             penalty-max: 8
             ignore-completion-date: True
         "6":
             github-prefix: os-task5
+            short-name: ЛР6
             taskid-max: 30
             penalty-max: 7
